+++ balancer.go
@@ -2,14 +2,9 @@ package main
 
 import (
 	"strconv"
-	"sync"
 )
 
-/*
-load balancer: for now, I've just implemented weighted round robin algorithms.
-it's borrowed from Nginx:
-https://github.com/nginx/nginx/commit/52327e0627f49dbda1e8db695e63a4b0af4448b1
-*/
+// load balancer: return which backend should we proxy to
 
 // Backend is the backend server, usally a app server like: gunicorn+flask
 type Backend struct {
@@ -27,61 +22,4 @@ func (b Backend) ToURL() string {
 // proxy.
 type Balancer interface {
 	Select() (*Backend, bool)
-}
-
-// WRR is weighted round robin algorithm
-type WRR struct {
-	lock sync.Mutex
-
-	upstream    []Backend
-	totalWeight int
-	weights     []int
-}
-
-// NewWRR return a instance with initialized weights & totalWeight
-func NewWRR(backends ...Backend) *WRR {
-	totalWeight := 0
-	weights := make([]int, len(backends))
-
-	for _, b := range backends {
-		totalWeight += b.Weight
-	}
-
-	return &WRR{upstream: backends, totalWeight: totalWeight, weights: weights}
-}
-
-// Select return the backend we should proxy
-// for example, weights of [5, 1, 1] should generate sequence of index:
-// [1, 1, 2, 1, 3, 1, 1]
-func (w *WRR) Select() (b *Backend, found bool) {
-	w.lock.Lock()
-
-	totalWeight := w.totalWeight
-	upstream := w.upstream
-	weights := w.weights
-	biggest := -1
-	biggestWeight := 0
-
-	for i := range weights {
-		weights[i] += upstream[i].Weight
-
-		if weights[i] > biggestWeight {
-			biggestWeight = weights[i]
-			biggest = i
-		}
-	}
-
-	if biggest >= 0 && biggest < len(weights) {
-		weights[biggest] -= totalWeight
-
-		// defer is too slow...
-		w.lock.Unlock()
-
-		return &w.upstream[biggest], true
-	}
-
-	// defer is too slow...
-	w.lock.Unlock()
-
-	return nil, false
 }
